Refactor task disabling logic and update property access
* Split `DisableTasks.kt` into platform-specific utility files: `AndroidDisableTasks.kt`, `KmpDisableTasks.kt`, `JvmDisableTasks.kt`, and `DisableTasksCore.kt`
* Update property accessors to use `scaffoldProperties()` for consistency
* Add `getBundleDependenciesOrNull` helper to `VersionCatalog.kt`
* Remove redundant `configureCommonAndroid` extension functions from `Extensions.kt`
