+++ plugins/src/main/kotlin/io/github/thomaskioko/gradle/plugins/utils/Extensions.kt
@@ -8,6 +8,8 @@ import com.android.build.api.dsl.LibraryExtension
 import com.android.build.api.variant.AndroidComponentsExtension
 import io.github.thomaskioko.gradle.plugins.extensions.AndroidExtension
 import io.github.thomaskioko.gradle.plugins.extensions.BaseExtension
+import io.github.thomaskioko.gradle.plugins.properties.PropertyKeys
+import io.github.thomaskioko.gradle.plugins.properties.scaffoldProperties
 import org.gradle.api.Project
 import org.gradle.api.plugins.JavaPluginExtension
 import org.gradle.api.provider.Provider
@@ -92,27 +94,6 @@ internal fun Project.androidMultiplatform(block: KotlinMultiplatformAndroidLibra
 }
 
 /**
- * Configures common Android settings shared across all Android project types:
- * namespace, compileSdk, minSdk, and lint.
- */
-internal fun Project.configureCommonAndroid() {
-    android {
-        namespace = pathBasedAndroidNamespace()
-        compileSdk = getVersion("android-compile").toInt()
-        defaultConfig.minSdk = getVersion("android-min").toInt()
-        lint.configure(project)
-    }
-}
-
-@Suppress("UnstableApiUsage")
-internal fun KotlinMultiplatformAndroidLibraryTarget.configureCommonAndroid(project: Project) {
-    namespace = project.pathBasedAndroidNamespace()
-    compileSdk = project.getVersion("android-compile").toInt()
-    minSdk = project.getVersion("android-min").toInt()
-    lint.configure(project)
-}
-
-/**
  * This function provides a concise way to customize the Android build process, including variant configuration and artifact management.
  */
 internal fun Project.androidComponents(block: AndroidComponentsExtension<*, *, *>.() -> Unit) {
@@ -184,9 +165,12 @@ internal fun KotlinMultiplatformAndroidLibraryTarget.jvmCompilerOptions(block: K
  * Throws an error if the package name is missing or empty in the Gradle properties.
  */
 internal fun Project.getPackageNameProvider(): Provider<String> =
-    stringProperty("package.name").map {
+    scaffoldProperties().packageName.map {
         it.takeIf { it.isNotBlank() }
-            ?: error("Required property 'package.name' is missing or empty in gradle.properties. Add: package.name=com.yourcompany.yourapp")
+            ?: error(
+                "Required property '${PropertyKeys.PACKAGE_NAME}' is missing or empty in gradle.properties. " +
+                    "Add: ${PropertyKeys.PACKAGE_NAME}=com.yourcompany.yourapp",
+            )
     }
 
 /**
